rename bottom tab navigator files
