+++ packages/app/navigation/bottom-tab-navigator.tsx
@@ -1,25 +1,15 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { useRouter } from 'next/router';
 
 import { TabBarIcon } from 'app/navigation/tab-bar-icon';
-import type { NextNavigationProps } from 'app/navigation/types';
+
+import HomeNavigator from 'app/pages/home';
+import PlaylistsNavigator from 'app/pages/playlists';
+import ProfileNavigator from 'app/pages/profile';
 
 const BottomTab = createBottomTabNavigator();
 
-export function BottomTabNavigator({
-  Component,
-  pageProps
-}: NextNavigationProps) {
-  const router = useRouter();
-
-  const component = useCallback(
-    (props) => {
-      return <Component {...pageProps} {...props} />;
-    },
-    [Component, pageProps]
-  );
-
+export function BottomTabNavigator() {
   return (
     <BottomTab.Navigator
       initialRouteName="home"
@@ -39,42 +29,21 @@ export function BottomTabNavigator({
     >
       <BottomTab.Screen
         name="home"
-        component={component}
-        listeners={{
-          tabPress: (e) => {
-            router?.push({ pathname: `/home` }, `/home`, {
-              shallow: true
-            });
-          }
-        }}
+        component={HomeNavigator}
         options={{
           tabBarIcon: ({ color }) => <TabBarIcon name="home" color={color} />
         }}
       />
       <BottomTab.Screen
         name="playlists"
-        component={component}
-        listeners={{
-          tabPress: (e) => {
-            router?.push({ pathname: `/playlists` }, `/playlists`, {
-              shallow: true
-            });
-          }
-        }}
+        component={PlaylistsNavigator}
         options={{
           tabBarIcon: ({ color }) => <TabBarIcon name="disc" color={color} />
         }}
       />
       <BottomTab.Screen
         name="profile"
-        component={component}
-        listeners={{
-          tabPress: (e) => {
-            router?.push({ pathname: `/profile` }, `/profile`, {
-              shallow: true
-            });
-          }
-        }}
+        component={ProfileNavigator}
         options={{
           tabBarIcon: ({ color }) => <TabBarIcon name="smile" color={color} />
         }}
